Organize tests project
