+++ package.json
@@ -19,13 +19,13 @@
 		"esbuild": "^0.20.0",
 		"jest": "^29.7.0",
 		"ts-jest": "^29.1.2",
+		"ts-node": "^10.9.2",
 		"typescript": "^5.4.3"
 	},
 	"dependencies": {
 		"axios": "^1.6.7",
 		"bad-words": "^3.0.4",
 		"form-data": "^4.0.0",
-		"ts-node": "^10.9.2",
 		"winston": "^3.11.0",
 		"winston-daily-rotate-file": "^5.0.0"
 	}
