+++ tests/unit/Utility.test.ts
@@ -1,6 +1,6 @@
 import "../helper/CfxGlobals";
 import "../helper/MockNatives";
-import { Utility } from "../../util/Utility";
+import { Utility } from "../../src/util/Utility";
 
 describe("Utility", () => {
 	it("should have the correct number of degrees in one radian", () => {
